Move Ui.java to ui package and modify relevant files
