+++ src/main/java/performative/Performative.java
@@ -3,6 +3,7 @@ package performative;
 // Java standard library imports
 import performative.parser.Parser;
 import performative.storage.Storage;
+import performative.ui.Ui;
 
 import java.io.IOException;
 import java.util.Scanner;
